Added nearest neighbor search and dynamic palette coloring
Nearest neighbor search is used to determine opposing units in range to
attack, and adjacent allied units to interact with.
Palette coloring refers to the act of applying a different color scheme
to existing sprite sheets.
Also added tile status windows, which show the name of the tile the
cursor is currently on and some of their stats.
Also added automatic UI snapping, s.t. user action doesn't overlap with
the UI.