Mainly fixed font formatting, and added some battle code.
There was an issue that the font had unnecessary spacing between each
letter, due to sprite sheet constraints, of every letter needing to be
bounded by the same rectangle. Now it dynamically adjusts the width of
each letter as to minimize unnecessary spacing.
Added code to properly load the UV unwraps of battle platforms.